requirements.txt
aiohttp==1.0.5 
async-timeout==1.2.1 
chardet==3.0.2 
discord.py==0.16.7 
multidict==2.1.4 
websockets==3.3
mpmath==0.19 
sympy==1.0
canvasvg==1.0.5
